# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/recommendations": {
      "post": {
        "operationId": "recommendations_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecommendationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Files a recommendation. **Scouts only** (§1.5).\n\nRecommending is the one action the whole reputation system measures, and it\nis measured per scout: `ScoutStats`, the §1.5 level tiers and the §1.5.1\nharmonic credibility all key off a scout identity. A coach or manager filing\none would accrue a reputation nothing in the product surfaces, and an\nacademy recommending a player to itself is not a signal at all.\n\nEnforced on the acting role, so a scout who is also a coach must be acting\nas a scout — see JwtStrategy.validate.",
        "tags": ["recommendations"]
      }
    },
    "/api/v1/recommendations/mine": {
      "get": {
        "operationId": "recommendations_listMine",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["recommendations"]
      }
    },
    "/api/v1/recommendations/inbox/awaiting-review-count": {
      "get": {
        "operationId": "recommendations_inboxAwaitingReviewCount",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Players in this manager's inbox that no coach has been asked about yet.\n\nDeclared before `academy/:academyId`, since Nest matches in declaration\norder and would otherwise read \"inbox\" as an academy id.",
        "tags": ["recommendations"]
      }
    },
    "/api/v1/recommendations/academy/{academyId}": {
      "get": {
        "operationId": "recommendations_listForAcademy",
        "parameters": [
          {
            "name": "academyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["recommendations"]
      }
    },
    "/api/v1/recommendations/academy/{academyId}/ranked": {
      "get": {
        "operationId": "recommendations_listRankedForAcademy",
        "parameters": [
          {
            "name": "academyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Manager-only: same inbox ranked by per-academy credibility (1.5.1/1.5.2).",
        "tags": ["recommendations"]
      }
    },
    "/api/v1/recommendations/academy/{academyId}/history": {
      "get": {
        "operationId": "recommendations_listHistoryForAcademy",
        "parameters": [
          {
            "name": "academyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "What this academy already settled — invited or turned down.",
        "tags": ["recommendations"]
      }
    },
    "/api/v1/recommendations/player/{playerId}/mine": {
      "get": {
        "operationId": "recommendations_myRecommendationFor",
        "parameters": [
          {
            "name": "playerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "This scout's own recommendation for a player — one per player (§1.5).",
        "tags": ["recommendations"]
      }
    },
    "/api/v1/recommendations/player/{playerId}/academy-state": {
      "get": {
        "operationId": "recommendations_academyStateFor",
        "parameters": [
          {
            "name": "playerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Where a player stands with the academy this manager runs.",
        "tags": ["recommendations"]
      }
    },
    "/api/v1/recommendations/players/{playerId}/review": {
      "post": {
        "operationId": "recommendations_assignReview",
        "parameters": [
          {
            "name": "playerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["recommendations"]
      }
    },
    "/api/v1/recommendations/player/{playerId}/my-review": {
      "get": {
        "operationId": "recommendations_myReviewFor",
        "parameters": [
          {
            "name": "playerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "This coach's review of one player, or null if nobody gave them that player.\n\nLets a coach answer from the profile they are reading rather than hunting\nthe same person down in their queue. Null is the rule, not an empty state:\na coach may only judge players an academy assigned to them.",
        "tags": ["recommendations"]
      }
    },
    "/api/v1/recommendations/reviews/mine": {
      "get": {
        "operationId": "recommendations_listMyReviews",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "A coach's own queue. Declared before `:id` — Nest matches in order.",
        "tags": ["recommendations"]
      }
    },
    "/api/v1/recommendations/reviews/{reviewId}/decision": {
      "post": {
        "operationId": "recommendations_decideReview",
        "parameters": [
          {
            "name": "reviewId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecisionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The coach's verdict, and the ratings that become the player's credible ones.",
        "tags": ["recommendations"]
      }
    },
    "/api/v1/recommendations/players/{playerId}/invite": {
      "post": {
        "operationId": "recommendations_invitePlayer",
        "parameters": [
          {
            "name": "playerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitePlayerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The manager invites an approved player, with a note they will read.",
        "tags": ["recommendations"]
      }
    },
    "/api/v1/recommendations/{id}/status": {
      "patch": {
        "operationId": "recommendations_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRecommendationStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["recommendations"]
      }
    },
    "/api/v1/recommendations/player/{playerId}": {
      "get": {
        "operationId": "recommendations_playerSummary",
        "parameters": [
          {
            "name": "playerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "A player's recommendation record: who vouched for them, with what weight, and\nthe public `globalWeight`. Per-academy extras stay in the inbox.",
        "tags": ["recommendations"]
      }
    },
    "/api/v1/recommendations/endorsing-academies": {
      "get": {
        "operationId": "recommendations_endorsingAcademies",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Academies that currently endorse me — the only valid targets for a SPECIFIC\nrecommendation. Drives the picker, so it can only ever offer valid choices.",
        "tags": ["recommendations"]
      }
    },
    "/api/v1/recommendations/scout-stats/me": {
      "get": {
        "operationId": "recommendations_myStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["recommendations"]
      }
    },
    "/api/v1/recommendations/scouts/{scoutId}": {
      "get": {
        "operationId": "recommendations_scoutProfile",
        "parameters": [
          {
            "name": "scoutId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "One scout's reputation page.\n\nVisible to players and academies, and refused to coaches — a coach answers\nfor the player in front of them, and knowing which scout is asking is\npressure rather than information. See `getScoutProfile` for the full\nreasoning; the rule is enforced there so a direct request is refused too,\nnot merely un-linked in the UI.",
        "tags": ["recommendations"]
      }
    },
...
  },
  "info": {
    "title": "FotSpot API",
...
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "The `accessToken` returned by any `/auth/*` login endpoint."
      }
    },
    "schemas": {
...
      "RecommendationType": {
        "type": "string",
        "enum": ["GLOBAL", "SPECIFIC"]
      },
      "CreateRecommendationDto": {
        "type": "object",
        "properties": {
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RecommendationType"
              }
            ]
          },
          "academyIds": {
            "description": "Required for SPECIFIC, rejected for GLOBAL — a cross-field rule, so it is\nenforced in the service alongside the endorsement check rather than split\nacross two places.\n\nCapped at 5: a \"specific\" recommendation addressed to every academy on the\nplatform is a global one wearing a disguise.",
            "minItems": 1,
            "maxItems": 5,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "playerId": {
            "type": "string"
          },
          "note": {
            "type": "string"
          }
        },
        "required": ["type", "playerId"]
      },
      "AssignReviewDto": {
        "type": "object",
        "properties": {
          "coachUserId": {
            "type": "string"
          }
        }
      },
      "ReviewDecisionDto": {
        "type": "object",
        "properties": {
          "decision": {
            "type": "string",
            "enum": ["APPROVED", "REJECTED"]
          },
          "note": {
            "type": "string",
            "maxLength": 1000
          }
        },
        "required": ["decision"]
      },
      "InvitePlayerDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "description": "When to come. The trial is created for this moment."
          },
          "location": {
            "type": "string",
            "description": "Where to come.",
            "minLength": 1,
            "maxLength": 200
          },
          "note": {
            "type": "string",
            "description": "What to bring, and anything else the family needs — the invitation itself.",
            "minLength": 1,
            "maxLength": 1000
          }
        },
        "required": ["date", "location", "note"]
      },
      "UpdateRecommendationStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["REVIEWING", "ACCEPTED", "REJECTED"]
          },
          "academyId": {
            "type": "string",
            "description": "Which academy is deciding. Required when the caller manages more than one, so\na verdict is never written to the wrong academy's target row."
          }
        },
        "required": ["status"]
      },
...
    }
  },
...
}